requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
boto==2.48.0
brotlipy==0.7.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
Collectfast==0.6.0
confusable-homoglyphs==2.0.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0
django-allauth==0.34.0
django-appconf==1.0.2
django-blog-zinnia==0.19
django-classy-tags==0.8.0
django-cleanup==1.0.1
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-easy-maps==0.9.3
django-embed-video==1.1.2
django-haystack==2.6.1
django-js-asset==0.1.1
django-machina==0.5.6
django-model-utils==2.0.3
django-mptt==0.9.0
django-newsletter==0.7
django-quiz-app==0.5.1
django-redis-cache==1.7.1
django-registration==2.3
django-smartfields==1.0.9
django-stdfile==0.5
django-storages==1.6.5
django-storages-redux==1.3.3
django-tagging==0.4.6
django-treebeard==4.2.0
django-widget-tweaks==1.4.1
django-xmlrpc==0.1.8
geopy==1.11.0
glob2==0.6
gunicorn==19.7.1
idna==2.6
ldif3==3.1.1
lxml==4.1.1
markdown2==2.3.5
mistune==0.8.3
mots-vides==2015.5.11
oauthlib==2.0.6
olefile==0.44
packaging==16.8
Pillow==4.3.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.18
pyparsing==2.2.0
python-card-me==0.9.3
python-dateutil==2.6.1
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
regex==2017.12.9
requests==2.18.4
requests-oauthlib==0.8.0
sendgrid==5.3.0
six==1.11.0
sorl-thumbnail==12.4.1
surlex==0.2.0
tox==2.9.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
Wand==0.4.4
whitenoise==3.3.1
Whoosh==2.7.4
